recyclerViewItem package 삭제 -> 리사이클러뷰 사용하는 액티비티 클래스 파일과 리사이클러뷰 아이템, 어댑터, 뷰홀더 하나의 클래스로 통합
