+++ README.md
@@ -7,8 +7,8 @@ TeamProject 2
     * e.g. 'copyright'로 파일 이름이 시작하면 xml(액티비티 아님)에 쓰인 이미지 파일 -> copyright_image.png
 
 ## project 관리
-* 패키지 세분화: myclass, mbti, account, recyclerview_item, etc.
-* 리사이클러뷰 관련 클래스(adapter, viewholder, decoration .. etc 는 개별 패키지로 관리하지 않고, 편의성을 위해서 임시로 activity와 함께 관리)
+* 패키지 세분화: myclass, mbti, account ... etc.
+* 리사이클러뷰 관련 클래스(recyclerViewItem, recyclerViewAdapter, viewHolder, decoration .. etc 는 개별 패키지로 관리하지 않고, 편의성을 위해서 activity와 함께 관리)
 
 ## Open source used
 - circle imageview: https://github.com/hdodenhof/CircleImageView
@@ -29,6 +29,7 @@ TeamProject 2
 - okhttp: https://github.com/square/okhttp
 - Motion Toast: https://github.com/Spikeysanju/MotionToast
 - google gson: https://github.com/google/gson
+- Groupie: https://github.com/lisawray/groupie
 <!-- ## Get Started Activity -->
 <!-- ![image](https://user-images.githubusercontent.com/43941383/114131739-1ab4a300-993e-11eb-8497-7c7785d186e0.png)-->
