--- app/src/main/java/com/example/mtmimyeon_gitmi/item/ItemCampusPhoneNumber.kt
@@ -1,8 +1,0 @@
-package com.example.mtmimyeon_gitmi.item
-
-data class ItemCampusPhoneNumber(
-    val name: String,
-    val location: String,
-    val siteUrl: String,
-    val phoneNumber: String
-)
