--- app/src/main/java/com/example/mtmimyeon_gitmi/item/ItemChat.kt
@@ -1,4 +1,0 @@
-package com.example.mtmimyeon_gitmi.item
-
-// 사실 imgUrl은 저장할 필요가 없으나, 통일성을 위해서 일단 추가
-data class ItemChat(val imgUrl: String, val name: String, val lastChat: String)
