--- app/src/main/java/com/example/mtmimyeon_gitmi/item/ItemMjuSite.kt
@@ -1,3 +1,0 @@
-package com.example.mtmimyeon_gitmi.item
-
-data class ItemMjuSite(val mjuImage: Int, val mjuText: String)
