--- app/src/main/java/com/example/mtmimyeon_gitmi/item/ItemSubjectBulletinBoard.kt
@@ -1,10 +1,0 @@
-package com.example.mtmimyeon_gitmi.item
-
-data class ItemSubjectBulletinBoard(
-    val idx: Int,
-    val title: String,
-    val content: String,
-    val date: String,
-    val writer: String,
-    val chatNum: String
-)
