--- app/src/main/java/com/example/mtmimyeon_gitmi/item/ItemSubjectBulletinBoardComment.kt
@@ -1,10 +1,0 @@
-package com.example.mtmimyeon_gitmi.item
-
-// 과목별 게시판, 게시글 댓글 유저 정보
-data class ItemSubjectBulletinBoardComment(
-    val idx: Int,
-    val name: String,
-    val commentContent: String,
-    val profileImgUrl: String,
-    var date: String
-)
